mrmap/requirements.txt
Django~=3.1.7
django-autocomplete-light~=3.8.2
django-formtools~=2.2
django-extensions~=3.1.1
django-simple-captcha~=0.5.14
djangorestframework~=3.12.4
django-bootstrap4~=2.3.1
django_bootstrap_swt~=2.0.1
django-fontawesome-5~=1.0.18
django-tables2~=2.3.4
tablib~=3.0.0
django-filter~=2.4.0
django-query-parameters~=0.2.3
xmltodict~=0.12.0
django-mathfilters~=1.0.0
django-leaflet~=0.27.1
django-mptt~=0.12.0
django-debug-toolbar~=3.2
cryptography~=3.4.7
lxml~=4.6.3
requests~=2.25.1
pycurl~=7.43.0.6
celery~=5.0.5
django-celery-beat~=2.2.0
django-celery-results~=2.0.1
psycopg2~=2.8.6
django-redis~=4.12.1
argon2-cffi~=20.1.0
django-nose~=1.4.7
model-bakery~=1.2.1
coverage~=5.5
sphinx_rtd_theme~=0.5.1

channels~=3.0.3
channels-redis~=3.2.0

django-guardian~=2.3.0
django-crum~=0.7.9